Move homepage order by into setting dialog
This is the first part of an overhaul in how per-page settings are
employed. More settings will be added and their storage will be switched
from cookie/localstorage to the database. The hole page will then be
done allowing users to switch to tabbed layout from the hole page and
allowing show whitespace to be persistent.

The CSS used in these new forms is quite modern (mainly :has()) but
hopefully most people have that by now (Can I use says it's baseline
2023), but either way a broken optional settings form isn't the end of
the world.