Poker: high card/flush shouldn't be sorted
Due to the way slices work in Go, even sub-slices, passing the shuffled
slice of cards to `straightCheck`, which then did an in-place sort to
calculate whether it was a straight or not, affected the original slice!

This meant people could depend on high card and flush to always be in
ascending order. Fix this with an explicit copy.