Save homepage settings to the DB
Create a new JSON column to store all per-page and global settings for a
given golfer. So the TOML file defines the settings and the DB just
stores an opaque JSON map.

Future commits will move columns from the users table to this new
system.

And then it should be nice and generic to show the correct per-page
settings of a given page and show all the settings on the settings page.

As part of this commit rename index to home so the name of the
route/template is consistent with the name of the settings.