Fix golfer settings checkbox layout
A previous commit flipped the order of checkboxes and labels in CSS. It
did this because it made the logic for auto-laying out forms a LOT
easier. While you could argue that it should've made the CSS specific to
only the currently auto-layed out forms given the entire settings page
is moving that way this is a quick fix to restore the old layout.