When determining the date of a previous solution for the discord bot, filter by scoring type.
Otherwise, a situation can arise where, for example, an earlier chars solution is replaced by a different one, with a lower byte count, but the chars solution keeps its date, because the char count is the same. Then the discord bot can erroneously report the earlier date for the bytes solution, unless we filter by scoring when doing the query.