discord bot: combine messages, even if the server is restarted. (#1169)
Store the last announcement state in the database.
Before modifying the last announcement, check whether it was last message in the channel. The previous method of listening for new messages wouldn't be as reliable now that this state is persisted, because a message could be created when the server is offline.