Add a tested time to solutions
This is a timestamp of when the solution was last tested.

Now that complete re-tests are taking longer and longer this will allow
a future commit to prioritise re-tests of solutions that haven't been
tested in a while and allow for partial re-tests without losing
"progress". This will also help for automated re-tests.

The column is updated both when the golfer runs the solution and when
the admin re-testing route does.

For now the column is nullable while we backfill, this will change in a
future commit once every solution has a last tested date.

Finally display the value on the public solution route.