Add random cases to brainfuck (#1175)
* Add random cases to brainfuck

As per discussion in the Discord, adds 5 semi-randomized test-cases to discourage hardcoding. Keeping in line with the hardcoded testcases, all randomized tests are printable ASCII and end with a newline.

* remove duplicate function

* remove unneeded import

* format

* Fixed case to print all ASCII

* curse you gofumpt

* why

* Restrict ASCII range

Restricts the range of ASCII generated by random cases to preserve legacy solutions that relied on the limited character range.