Use outputTests() in more holes
This both simplifies the code and in the longer term will allow us to
change the return type of the function to [][]test and de-couple the
definition of a hole and the running of a solution. This will allow the
holes to live in a different package like config so they can be attached
to the Hole struct they belong to.