First pass at /wiki MathJax support
It's far from perfect and plenty of maths on /wiki/spigot-algorithms
doesn't yet render but it's a start.

The font support is incredibly hacky as it's not running through
esbuild. There's a Makefile target to copy the fonts out of node_modules
that'll need to be ran on every update.