Show consistent code blocks on Asm wiki page
Because the Asm wiki page has code blocks in both AT&T and Intel syntax
it uses a combination of tagged and untagged fenced code blocks.

Becuase we previously skipped lang-asm blocks we ended up with a mixture
of native and CodeMirror code blocks which has varying font sizes and
line heights and just looked a bit shit. A better fix is to not skip the
conversion of asm code blocks but to just make them untagged plain text,
this commit does just that.