Zeckendorf Representation: Limit numbers to [1, 2^31).
For some languages, it's much more convenient if the arguments can be represented using a signed 32 bit integer (previously the number required using an signed 64 bit integer).
I don't think this change would significantly affect the way that a golfer would approach the problem.
This change also keeps the output line length to around 150 bytes, instead of over 300 bytes, which is more manageable in difference scenarios, like reading a diff.