Add /stats/unsolved-holes
This page shows hole/lang combinations that have yet to be solved,
"easy" diamonds/unicorns if you will.

It doesn't account for the unlikely scenario where a combo has just a
bytes solution or just a chars solution but instead just shows a link if
there's zero solutions of any type for that combo.

Closes #986