Stop blocking hole buttons after popup
Because the popups container had padding and gap set and because the
popups inside where only toggled with opacity the container still had
physical size after the popups faded.

Fix this by switching to visibility collapse which unlike display none
can be animated. However visibility collapse doesn't work with grid so
we must switch to flex. Because of this we can't have our two column
approach where the first column is just padding, instead we must set a
sensible width for all popups. This does have the nice side effect that
every popup is always the same width now though.