Group by lang on golfer feed
Conceptually similar to the existing group by cheeov logic. I went with
grouping by lang reather than grouping by hole because it seems to be
more common to solve a bunch of holes in a few languages rather than the
other way around so this way eliminates the most noise.

The limit is reduced from 100 to 75 to match the per page elsewhere but
in practice we often get more lines than we did before because the limit
isn't passed down to the SQL anymore, we keep going until we get 75
grouped lines (or we run out of data).