Standardise use of Python in langs
Build it from source indentically to the Python image. Similiar in
spirit to the Rust-based langs.

Hy gets a lot bigger because we now have the .pyc files. A lot won't be
needed but removing them all slows down startup significantly. Longer
term we should find more libraries we can cull.

Also simplify the version command, it used to matter that we only
printed the version and not any other crap like the name of the language
but these days build-langs takes care of extracting just the version.

Finally, since we install the langs via pip it makes sense to use it as the
source of truth.