Correct localise last tested day on /admin
This was subtle but because we store solution times as timestamp rather
than timestampz to save space but always store them as UTC to be correct
we need to tell postgres it's in UTC before converting it the user's
timezone, in this case, Europe/London. Yay for BST making this bug
visible and me working after midnight ;-)

The DATE() call also must go on the outside, I don't know what I was
thinking. While here use a shorter name with less stutter.