Port Card Number and Pangram to test{}
This required a small tweak to outputTests() which I think is valid, we
skip over any empty strings in expected output. This allows these two
holes to have arguments that produce no output. I don't think any
current holes support empty string as an expected output for an arg.

The stutter in Card Number of repeating the valid hardcoded test cases
is a little unforunate but it's better than having yet another struct we
turn into test{}. Maybe it points to having a simple builder func that
takes an arg and a bool.

We now unconditionally print the arg separator in outputTests() which is
fine since we merged #1333 which unconditionally strips trailing
whitespace from every hole's expected output.