Generalise langs used svg embedding
This generalises how we embed only the SVGs of the used languages on the
page in order to save page size. This represents an optimisation for
/recent/golfers which previously contained the SVGs of every language.

The new template functions could be more generic using interfaces but
since we only have one type for now let's make them concrete and more
efficient for now.

The hasField template function might be useful in other templates.

This also allows us to signicantly simplify routes.homeGET as it now
only needs the cards as its data.