Add a "setting" template func
This solves a use case where you want the value from a setting and the
golfer may or may not be logged in. If they're not we should return the
default value for the given setting. This fixes a blank page on the
golfer cheevo page if you weren't logged in.