Make "Rockstar 2" live
Omit an entry in feeds as the plan is to have this language replace
Rockstar and then inherit the name.

A future commit will display a banner to anyone with a R1 solve in a
hole without a corresponding R2 solve telling them to port their
solution as R1 will be sunsetted.

What to do with remaining R1 solutions at that point is up for grabs,
one option might be to migrate them to notes so authors can still view
them.