Tighten up Clojure implicit output prevention (#1759)
It was possible to add a comment to the end of Clojure code and
re-enable implicit output.

I haven't found a way to actually exploit this in any hole to save bytes,
since the implicit output is printed as if with `pr`, so eg strings are
quoted. I don't think it can be used to help with quine, for example.

I discovered it when I accidentally left a comment at the end of my code,
and I was confused.

This adds a newline to render `;` comments ineffective, and adds an
additional `(print)` so that `#_` will only comment out the first form.