Serve lang logos externally
Rather than embedding the SVg of a lang logo directly into the page and
serving it on every request, reference it from the page like so:

  <svg><use href="/dist/svg/perl-ZV53OFZM.svg#perl"/></svg>

This way it'll be cached and only fetched once. The use of <use> as
opposed to <img> is so that it inherits currentColor and reacts to hover
and light/dark mode.

We also aggressively compress them with brotli/zopfli for live now that
they're no longer embedded inside dynamic pages.

I suspect it makes sense to do this for all SVG files eventually but
given they need IDs (which all logos do) and lang logos are generally
much larger, just do lang logos for now.

This replaces the work in b89cc77 as it not longer matters how many
times we reference the same lang logo on a page now as it's only fetched
up to once.

Also add the logo URL to the hole JSON for the frontend hole UI and
therefore document it on the API.