Rename "Placeholder" to "Tutorial"
I never really liked the name Placeholder, it implied (to me) that it
was a placeholder hole that is temporary. So instead use another popular
suggestion from the hole's lengthy issue conversation, Tutorial.

While here flesh out the description a little more, and refer to the
code as sample code, not placeholder as that's what the UI does (another
reason to dislike the name placeholder). Incidentally langs.toml refers
to this code as example not sample, we probably ought to be consistent.

After amending the enum to include the new value (you can rename enum
values but not move them), existed data was migrated with:

  BEGIN;
    UPDATE authors         SET hole = 'tutorial' WHERE hole = 'placeholder';
    UPDATE discord_records SET hole = 'tutorial' WHERE hole = 'placeholder';
    UPDATE notes           SET hole = 'tutorial' WHERE hole = 'placeholder';
    UPDATE solutions       SET hole = 'tutorial' WHERE hole = 'placeholder';
    UPDATE solutions_log   SET hole = 'tutorial' WHERE hole = 'placeholder';
  COMMIT;

  REFRESH MATERIALIZED VIEW medals;
  REFRESH MATERIALIZED VIEW rankings;
  REFRESH MATERIALIZED VIEW points;