latest-langs: fix Erlang/Julia "latest" stable
Rather than sorting by date, sort by actual version number.

In theory we could move all the GitHub URLs to the API but the
rate-limiting is pretty harsh, only 60 reqs an hour with a key.
Longer term we could re-design latest-langs to look up multiple
languages at once with a single GraphQL query or something but this is a
quick fix to get these two languages DTRT

Also do Odin so we can remove the special-case from the next block.