Replace NullCountry with *Country
This is simpler than before so long as we never have any invalid values
in the DB as that will panic. This and the previous commits have made it
so we no longer have an invalid empty string country/country_flag,
instead we have a two letter ID, or NULL.

The generated column has moved position as it required dropping and
re-adding to get the new definition, without the coalesce.