Delete config.Null*
And finally we can do what the last few refactorings have aimed to do,
drop these types. 99% of the time we can scan into a pointer or use a
(possibly nil) pointer to pass as SQL. The rare times we only have a
string ID we can use the generic null.Null[...] type which these
predate.