Calculate digest of each lang
This will be used to accurately determine if a solution needs to be
re-ran when a language changes. It's more sensitive than version as it's
literally the hash of all of the files that make up a language from all
docker layers.

Takes ~30s to calculate at build time on my laptop so it's only re-ran
when either a lang layer changes or the hasher itself does. This is
still quicker than building assets (mainly because zopfli) so it's not a
huge deal except unlike zopfli it is ran on dev startup. We could use
dummy values from dev but I'd rather not, especially while this feature
is in development. Choice of hash algo and go routines might help bring
this time down in future. As would shrinking existing langs ;-)

Written in Go because the stdlib makes it easy and we can just copy the
static binary to the scratch layer and run it at build time, very nice.

This will also be useful in determining when a solution runtime must
overwrite a saved value or we can keep the lowest value. The rough plan
is we'd have a hash of language and some kind of sentinal version for
both hole version and platform version. If they're they all match we can
take the lowest otherwise we take the latest and update the values.

Updates #592