Optimise & simplify Set hole
The previous implementation did a lot of work per iteration which is bad
when it could take many iterations to produce 100 test cases as there's
randomness involved. It also mutated the global deck which could cause
races. Instead of shuffling the deck we indexed randomly into the slice
(which is now an array since it never grows/shrinks). The string
builders weren't really doing anything and there's no need for a map,
just use rand.Perm instead to ensure no repeated cards.