Refactor hole functions
Instead of returning []Run use a new struct []Answer. This better
ecapsulates the name of what a hole should return, we shouldn't re-use a
Run which contains the result of running the code as well.

Along with this we move these new answer-returning funcs to be on the
Hole struct itself and remove most of the entries from the crazy long
switch statement.