Ruby thread size of 500MB is too big
With this live solutions in all these languages were dying with:

  failed to allocate memory (NoMemoryError) exit status 1

And we were still getting some solutions in iogii to die with:

  ERROR: Stack Overflow (You can increase the limit with the shell
  command: export RUBY_THREAD_VM_STACK_SIZE=<size in bytes> or
  compile to haskell with -hs to remove the limit).

So I suspect Haskell might be the way forward, either way we have to
remove this env value as it breaks non-experimental solutions. Compiling
with LTO under Clang seems harmless enough though.

Updates #1770