Bump Rust
Uiua couldn't be bumped because it can't be rebuilt atm due to an
unpinned dep of a dep on zip2 which is currently broken. See
https://github.com/zip-rs/zip2/issues/328

It's an optional feature of Uiua, not sure why the dep is fetched when
we're supposedly not building optional features.

uiua → calamine → zip2

This is also affecting the bump of Uiua itself to 0.15.0