Remove chdir from Prolog runner (for now)
At least one existing solution relied on being at the root directory
rather than inside /tmp, so maintain this behaviour for now.

Longer term we ought to be consistent, on one hand a pointless chdir is
a pointless chdir and / makes more sense than /tmp. However some
languages expect a writable working directory so maybe for consistency
it would be better to transition to /tmp everywhere and have people fix
any fallout. I guess if we do get to that point then there's no reason
the chdir couldn't be lifted into run-lang.c codebase instead but we're
not there yet.