Fix another bug in utils/update_sql_from_api.go
It's possible to have an asm solution whose byte count exceeds the
400KiB source code limit as that limit is only for source code. This
populate script tried to make dummy source code of the same length as
the byte count which isn't necessary for asm solutions, not doing so
avoids work and allows the script to complete.

It's all Sisyphus' fault with their 573,611 byte asm solution!
https://code.golf/golfers/sisyphus-ppcg/fibonacci/assembly/bytes