Make the rankings on the hole page scrollable
This basically ports the feature from the WIP multi-window hole page to
the normal one. In the process the markup and css was heavily refactored
to be simpler, instead of a complex grid structure we have two logical
flexboxes with the second being the rankings view and hidden on mobile
for space reasons.

This simpler markup opens the door to having the editor and the rankings
being any size, which is the main feature people seem to like about the
multi-window hole page.

This change was more work than it ought to be, the relationship between
the two pages, especially the CSS is a hot mess and it's very hard to
change one page without massively breaking assumptions on the other. In
the long term we really just need one view, in the short term though I
might just make the multi-window page a copy with no shared markup or
CSS.

The JS for jumping to the right row on refresh when on the "me" tab
needed a re-think as unlike multi-window mode, the regular hole page has
a body scrollbar and scrollIntoView would move that, which was very
jarring on code submission.