Build codegolf/lang-java from source (#2260)
* Build codegolf/lang-java from source

This commit is about building Temurin, the OpenJDK-based image of
Eclipse Adoptium, from source. It has advantages, but also a
disadvantage.

Currently Java, CJam, Groovy, Kotlin, Scala, and Vyxal use base images
with everything pre-built. Easy, but often very annoying. When a new
release is launched, it sometimes takes more than a week for the new
images to be pushed to Docker Hub. Completely understandable, because
there are many targets and architectures that need to be processed in a
certain order of priority. Docker is of course one of the last in line.
With this commit, that long wait can be reduced considerably by building
the image for Java itself and letting the other mentioned languages use
that image. That in turn results in images with much smaller file sizes.

The disadvantage is that it now takes time to build the image for Java,
but I think you get a lot more in return.

Closes #2221, closes #2222, closes #2223, closes #2224, closes #2225,
closes #2226

* Minor updates