Move the website to Go 1.25rc2
No need to bump the go.mod lang version just yet and break the linter as
we don't really need anything in the stdlib from the new version. It
would be nice to try out the new json/v2 package but that currently
requires setting the GOEXPERIMENT env var at build time which is a bit
of a faff for editors.