Test Kotlin exceptionally with an alternative hole
In general, Mahjong and Musical Chords, which are also full of Unicode
arguments, are always faster than Rock-paper-scissors-Spock-lizard. I
started with Mahjong because it only has two runs, but Musical Chords
has only one and is even faster than Mahjong. It might be wise to test
Kotlin with Musical Chords.