Simplify POST /solution with json/v2
Most of this is the tooling around feeding the Go experiment to the
various places. The route is slightly stricter in that you now need to
match the JSON key case, which the testing library wasn't but the
website JS was, third-party submitters might need updating if they
weren't already using lowercase names.