Bite the bullet, bump Zig
As per the changelog, this is an "extremely breaking" release as the
whole way you read/write to files has changed!

See https://ziglang.org/download/0.15.1/release-notes.html#Writergate

Zig has a history of making breaking changes like this, which is fair
for a 0.x release. In hindsight we probably shouldn't have marked it as
stable on code.golf yet and undoing that would break achievements like
Pangramglot as it's currently the only stable lang with a "z".

Either way, we have to upgrade, hopefully this is the last breaking
change for Zig 🤞

Closes #2370