Replace .headless class with selectors
It took me a while to remember what the class was used for, the new
selectors are both self-descriptive and don't require any special
annotation in the markup which will be handy for things like the wiki.

The mini scores table on the hole page is a special-case in that the
border is provided by the container, so it's right that it should have
to override the border property of the table imo.