Use CDS to create an archive that can be loaded at runtime
Currently, as before, a pre-built Temurin image from Eclipse Adoptium is
being used for this purpose. It works fine with that, but not yet with
the new build method. The "--enable-jvm-feature-cds" flag hasn't yielded
any positive results yet. I'll see if `strace` can provide more clarity
on what's being done later. In any case, the boot time is significantly
shorter with the use of Class Data Sharing.