Show progress steps on incomplete cheevos
Only enabled for a handful of cheevos for now but will be expanded to
any that make sense in the future, it'll be especially useful for
cheevos like Pangramglot and the upcoming Count to Ten cheevos where
it's not always clear which steps have been acomplished.

I especially like how it doesn't leak "hidden" steps for the Archivist
cheevo and will be extended to the upcoming Zoodiac Signs cheevo. While
nothing is that hidden as you can both look at other profiles that have
completed some of the steps or just look at the source code of the
website I still think this helps preserve the fun.