Some adjustments to the Hare language (#2448)
Initially, I worked on a way to remove Hare from the "build-langs"
script, and while that worked, one thing led to another. Here's a
summary:

* The "latest-langs" script was no longer correctly displaying the
latest release and was therefore modified to do so again.
* The Dockerfile entry was calling the actual binary in "/usr/local/bin"
instead of the modified runner, which was renamed as a result.
* Modifying the "latest-langs" script showed that a new release was
available, and Hare was updated accordingly.
* The final image size was reduced by removing symbols and sections from
the built binaries.
* Ultimately, Hare was removed from the "build-langs" script because the
issue with retrieving version information had apparently been resolved.

Now only the giant rabbit remains...