Rebuild BASIC from source on Debian "trixie" 13 (#2438)
* Build BASIC from source on Debian "trixie" 13

* Bootstraps FBC with only minimal features needed to compile another.
This is one of several workarounds for the missing "libtinfo5" package
in the "trixie" suite.
* Some headers that were likely unused have been removed. For example,
the "include" folder currently contains a folder of almost 15 MiB that
targets Windows.
* All contents of the folders installed by the build file, except for
the four deleted folders, now remain intact.
* The number of duplicates of many library files has been reduced to
zero.
* The final image size has been reduced from 200 MiB to 68 MiB.

* Addition of additional libraries to call GCC

This has been tested with three different invocations, and I think this
is it. Otherwise, I give up.