Save solution runtime to DB
For a given submission it'll be the longest of each of the runs but only
updated in the DB if it's lower than the existing, or the strokes
change, or the failing flag changes, or the lang digest changes. This
way you get to keep your fastest run all things being equal.

It should probably take into account some kind of hole digest for when
the hole changes and maybe some kind of platform version when hardware
or the running framework changes in such a way that would affect the
timing. Though in practice the lang digest changes frequently enough for
this not to matter much.

Future commits will make this metic available for all to see, it's no
more information leakage than we currently have via the test solution
feature and allows for some fun insights imo.

Secondly this will be an invaluable tool for discovering speed
regressions with languages as their container is rebuild, there's been a
few recent speed regressions that I would like to be able to measure and
catch.

The column will be made not null once all solutions are re-tested.

Updates #592