Remove zero runtime check from admin_solutions.go
This was added in the previous commit out of an abundance of caution but
it's actually wrong, there are cases where we can have zero runtime for
non-intermittent errors, when we return early without executing the code
because the code itself is wrong, usually for quine protection.

For example in our datbase we have a few Go quines that are saved and
marked as failing because they've not been updated since we forbade
embed in Go quines. Without this change we can't correctly set their
runtime to zero.