Add a Minesweeper test case with a cell of 8
This is the highest number you can get in Minesweeper, a cell surrounded
by mines and is an important edge-case to test. We probably ought to
have test cases for 6, 7, and 8 as well. I've replaced the previously
final test case to ensure solutions that hardcode an argc of 100 keep
working.

Updates #1025