Clean build on Debian "bullseye" 11
This has two problems:

* Neither Alpine nor more recent Debian versions can bypass the
instruction that uses a deprecated system call (`sysctl`).
* The question mark operators representing `QOut` (`?`) and `QQOut`
(`??`) cannot be used because neither symbol is found.

This last problem can be solved by adding `#define` instructions at the
top of the Harbour source file in the runner file, but I'll first
compare some files from the stable version (3.0.0) and the development
version (3.2.0) to see if I can make any sense of it.