Improve Java performance using Application Class Data Sharing
The first run creates the archive and remains virtually unchanged in
terms of speed. Subsequent runs load the archive. Because only Java is
currently using it, the default CDS archive is removed from the other
containers to save space. I clearly saw the difference yesterday.
Strange, you didn't. This time, however, it's not the same and should
yield better results.