Revert "Use code-golf/oauth in routes/callback"
This reverts commit 5dcb65d95e15ad0d943a2f8852f7e31d5b52ace1.

Log in /callback and golfer connections differ in RedirectURL so having
them share the same structs broke this logic. The logic in redner.go
where we take a shallow clone of the struct and adjust the redirect URL
to point to the current page is especially fragile and some OAuth
providers won't allow query params on the end of their callback URLs.

So we probably need to unify our approach and come up with a better way
to store the page to redirect back to than having to clone the struct to
edit it.

This blocks logging in with more than just GitHub as how it stands, the
GH app is hardcoded to go to /callback (and /callback/dev), and the
other three are hardcoded to go to /golfer/connect/{provider}.

Also the way we currently have it, I think /callback is subject to an
open redirect vulnerability as it blindly trusts whatever redirect_uri
is passed to it.

So I think one simple approach with a slight hit to UX would be to have
a shared route, say /oauth/{provider}, and just redirect to / if the
user wasn't already logged in, or redirect to /golfer/settings or
probably a new page /golfer/settings/oauth if they were logged in and we
were either adding a new connection or refreshing an existing one (I'd
like to add a refresh button).

The main hit would be the log in URL that appears on the hole page if
you're not logged in as it would cause people to lose the page they were
on once logging in.

Long story not very short, reverting this for now as it broke log in and
the path forward needs a lot of thinking about.

Also worth noting, we're currently not actually using OpenID Connect, or
at least we're not making use of the ID token, instead we're just making
a server side call to an API after oauth to get the identity. That would
be a nice improvement, assuming all our current providers support OIDC
and not just OAuth.