There are number of changes, one of which is _very_ important
By specifying `--with-lto=full`, we can build as before. This makes
"zlib" inaccessible without removing that file, as it won't be created.
This is likely because the "zlib-dev" package isn't being added, and I
suspect it's only needed for "ensurepip".

However, I forgot to remove the `binascii.cpython-*x86_64-linux-musl.so`
file, as the "base64" module, which imports the "binascii" module, could
still be imported. Now it can't.

Furthermore, the files from the first three removals won't be created
anymore. The container image will also be smaller, which is nice.