Remove /rankings/recent-holes
A sliding-window leaderboard of the ten most recent holes never quite
fit into the site very well imo. It was both this weird complex bolted
on feature and not flexible enough. So I'm killing it to make room for
something better in the future.

My very hand-wavey vision is more dynamic rankings pages that groups by
the type of hole, as well as maybe the date they were released, we could
group by holes of 2025, or category Art, or some combination.

There's also been chatter than some of the currently misc leaderboards
would be more useful if you could filter them like you can the top-level
leaderboards.

This all points to a more flexible approach to leaderboards, I think it
would be nice for a golfer to either embed their preferred leaderboards
on either a redesigned homepage and/or their profile, a bit like the
mini-leaderboard on the hole page it could be centered on them but allow
viewing the top records or how their friends are doing.

We might also want to consider redesigning categories before comitting
to leaderboards based on them, see #630 for switching to tags, and #1122
for adding cheevos for progression along the categories.

See also #550 which allowed sorting leaderboards by the headers, and the
various other issues floating about suggested new and improved
leaderboards.

tl;dr we need better leaderboards, /rankings/recent-holes was a stopgap
at best.