Speed up /rankings/holes/all/all/bytes query
Do this by expanding the points mat view to cover all the stuff we need,
it's only a little bit more work for the mat view but massively speeds
up the query on this page. Given this is the first rankings page people
land it on when they click on the rankings tab it makes sense to
optimise it, and it was getting rather slow.

It takes the page from ~150ms to ~40ms.