Start linking to avatars directly
And avoid /golfers/{name}/avatar, this saves a thundering hurd where a
browser with an empty cache hits /rankings then makes 75 requests for
avatars. Instead we can just select the avatar URL from the DB at the
same time as laying out the page. I mean it's still a thundering hurd
for sites like GitHub, but they're bigger than us :-P

We do this by creating a view across users and connections, ATM this is
hardcoded to only look at the GitHub connection but it's one place to
tweak in the future once we allow avatar choice/different oauth
providers to log in.

This initial commit only ports users of golfer-link.html to this new
paradigme as avatar use is prolific and the commit was already getting
unwieldy. Future uses will follow.