Introduce GolferLink struct, use it everywhere
This is the smallest unit of a golfer consisting of an avatar URL, a
name, and an optional country (name/flag). It's a pretty common use-case
to want this information in table listings throughout the site.

The medium fat Golfer struct is then built from that and the final
GolferInfo struct is built from Golfer. One slight footgun is Country vs
CountryFlag, the latter is the public country, I suspect it would be
good to rename that to something like CountryPublic to make it clear
it's the one to use in public-facing output, as people can hide their
country.