Erlang 101 1 11
This already simplifies the way Erlang scripts can be run, albeit
modestly. All solutions will likely remain intact for now by just
tinkering with the source file enough. This eliminates the compiler,
which reduces the container size somewhat. The Dockerfile also allows
for some movement, but that's for later. Now I want to test that '11'
mess on an Alpine distribution. If that works, Erlang is in trouble,
because I've already provisionally set the website's runtime to seven
minutes, if you know what I mean.