Avoid truncating multi-line command line arguments
That is to say: more or less. This fully displays multi-line command
line arguments, so some progress has been made. Unfortunately, this
isn't the solution yet, as it leads to the same problem as with Stax:
each line is treated as a separate argument. I believe executing an
empty source file should also fully display the first command line
argument. That would solve the problem.