Finally figured out how I'm going to tackle it
Before this commit, there weren't nearly enough words, but it's easier
to first determine all valid words in a given grid and then continue to
generate the list with this and that, and then it can fail as often as
it likes. So the dictionary will shrink again, but that's for the end.