Add "📺 Prime Time" cheevo
Thanks to @kg583 for the idea, @AlephSquirrel for the emoji, and @Yewzir
for the initial PR.

I've moved the cheevo from hole/lang specific to date specific as that's
how the cheevo was originally conceived and it just makes more sense.

I've also used literal numbers rather than named constants for the
months as it makes it clearer that we're testing all the prime months.

I've decided against adding the date ranges to cheevos.toml and
therefore displaying banners as I think it might be too noisy as there's
a lot of dates, but maybe that's worth revisiting in the future.

The cheevo was backfilled with the following SQL:

  insert into cheevos
       select min(submitted), user_id, 'prime-time'
         from solutions_log
        where hole in ('prime-numbers', 'prime-numbers-long')
          and date_part('month', submitted) in (2, 3, 4, 5, 11)
          and date_part('day',   submitted) in (2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31)
     group by user_id;

Closes #2548
Updates #2299

Co-authored-by: kg583 <kg583@users.noreply.github.com>
Co-authored-by: AlephSquirrel <AlephSquirrel@users.noreply.github.com>
Co-authored-by: Yewzir <Yewzir@users.noreply.github.com>