Prevent the "nibbles" folder from being copied during runtime
I had done this before in previous commits, but without much trouble.
Initially, I couldn't figure out why it kept failing, but that mystery
has since been solved. The reason it works for Gleam, for example, is
that a source file already exists due to the project creation during the
build phase. If the ownership and/or permissions are changed only after
the file is created, the compiler encounters no problems and can proceed
as planned. I've tried implementing this here, and it works. However, if
you run the `chmod` command before the file is created, it won't work.