Reset forms inside dialogs on dialog close
This commit extacts the common logic for clearing forms inside dialogs
that was used for both the settings modals and the search nav modal.

Technically it now runs on close rather than just before open but the
effect should be the same as every form starts clean. This refactor
allows further commits around dialogs.